Basic changes, added two pages, attempted to add a dropdown (unsuccessful and commented out code)
Made some basic changes to the Navbar, and the Avatar component (changed size).  Added a button to the profile page to create a new recipe from there.  Also added a basic createRecipe page.  Added borders to Avatars/profile pictures.  I attempted to create a dropdown for the avatar in the top right of the screen, but was unsuccessful so I commented out the code.